+++ src/meltano/cli/elt.py
@@ -5,11 +5,12 @@ from __future__ import annotations
 import logging
 import platform
 import typing as t
+import uuid
 from contextlib import asynccontextmanager, nullcontext, suppress
 from datetime import datetime, timezone
 
 import click
-from structlog import stdlib as structlog_stdlib
+import structlog
 
 from meltano.cli.params import (
     InstallPlugins,
@@ -18,6 +19,7 @@ from meltano.cli.params import (
     pass_project,
 )
 from meltano.cli.utils import CliEnvironmentBehavior, CliError, PartialInstrumentedCmd
+from meltano.core._state import StateStrategy
 from meltano.core.db import project_engine
 from meltano.core.elt_context import ELTContext, ELTContextBuilder
 from meltano.core.job import Job, JobFinder
@@ -33,9 +35,6 @@ from meltano.core.tracking.contexts import CliEvent, PluginsTrackingContext
 from meltano.core.utils import run_async
 
 if t.TYPE_CHECKING:
-    import uuid
-
-    import structlog
     from sqlalchemy.orm import Session
 
     from meltano.core.plugin.base import PluginDefinition
@@ -49,7 +48,7 @@ DUMPABLES = {
     "loader-config": (PluginType.LOADERS, "config"),
 }
 
-logger = structlog_stdlib.get_logger(__name__)
+logger = structlog.stdlib.get_logger(__name__)
 
 install, no_install, only_install = get_install_options(include_only_install=True)
 
@@ -114,6 +113,15 @@ class ELOptions:
         "--merge-state",
         is_flag=True,
         help="Merges state with that of previous runs.",
+        hidden=True,
+    )
+    state_strategy = click.option(
+        "--state-strategy",
+        # TODO: use click.Choice(StateStrategy) once we drop support for Python 3.9 and
+        # use click 8.2+ exclusively
+        type=click.Choice([strategy.value for strategy in StateStrategy]),
+        help="Strategy to use for state updates.",
+        default=StateStrategy.AUTO.value,
     )
     run_id = click.option(
         "--run-id",
@@ -140,6 +148,7 @@ class ELOptions:
 @ELOptions.state_id
 @ELOptions.force
 @ELOptions.merge_state
+@ELOptions.state_strategy
 @install
 @no_install
 @only_install
@@ -164,6 +173,7 @@ async def el(  # WPS408
     state_id: str | None,
     force: bool,
     merge_state: bool,
+    state_strategy: str,
     run_id: uuid.UUID | None,
     install_plugins: InstallPlugins,
 ) -> None:
@@ -194,6 +204,7 @@ async def el(  # WPS408
         state_id=state_id,
         force=force,
         merge_state=merge_state,
+        state_strategy=state_strategy,
         install_plugins=install_plugins,
         run_id=run_id,
     )
@@ -218,6 +229,7 @@ async def el(  # WPS408
 @ELOptions.state_id
 @ELOptions.force
 @ELOptions.merge_state
+@ELOptions.state_strategy
 @install
 @no_install
 @only_install
@@ -243,6 +255,7 @@ async def elt(  # WPS408
     state_id: str | None,
     force: bool,
     merge_state: bool,
+    state_strategy: str,
     install_plugins: InstallPlugins,
     run_id: uuid.UUID | None,
 ) -> None:
@@ -274,6 +287,7 @@ async def elt(  # WPS408
         state_id=state_id,
         force=force,
         merge_state=merge_state,
+        state_strategy=state_strategy,
         install_plugins=install_plugins,
         run_id=run_id,
     )
@@ -297,6 +311,7 @@ async def _run_el_command(
     state_id: str | None,
     force: bool,
     merge_state: bool,
+    state_strategy: str,
     install_plugins: InstallPlugins,
     run_id: uuid.UUID | None,
 ) -> None:
@@ -316,6 +331,15 @@ async def _run_el_command(
     transform = transform or "skip"
 
     select_filter = [*select, *(f"!{entity}" for entity in exclude)]
+
+    # Bind run_id at the start of the CLI entrypoint
+    run_id = run_id or uuid.uuid4()
+    structlog.contextvars.bind_contextvars(run_id=str(run_id))
+
+    _state_strategy = StateStrategy.from_cli_args(
+        merge_state=merge_state,
+        state_strategy=state_strategy,
+    )
 
     job = Job(
         job_name=state_id
@@ -341,7 +365,7 @@ async def _run_el_command(
             select_filter=select_filter,
             catalog=catalog,
             state=state,
-            merge_state=merge_state,
+            state_strategy=_state_strategy,
             run_id=run_id,
         )
 
@@ -377,13 +401,12 @@ def _elt_context_builder(
     dry_run: bool = False,
     full_refresh: bool = False,
     refresh_catalog: bool = False,
-    select_filter: list[str] | None = None,
+    select_filter: list[str],
     catalog: str | None = None,
     state: str | None = None,
-    merge_state: bool = False,
+    state_strategy: StateStrategy,
     run_id: uuid.UUID | None = None,
 ) -> ELTContextBuilder:
-    select_filter = select_filter or []
     transform_name = None
     if transform != "skip":
         transform_name = _find_transform_for_extractor(project, extractor)
@@ -402,7 +425,7 @@ def _elt_context_builder(
         .with_select_filter(select_filter)
         .with_catalog(catalog)
         .with_state(state)
-        .with_merge_state(merge_state=merge_state)
+        .with_state_strategy(state_strategy=state_strategy)
         .with_run_id(run_id)
     )
 
@@ -458,7 +481,8 @@ async def _run_job(
 
 @asynccontextmanager
 async def _redirect_output(
-    log: structlog.stdlib.BoundLogger, output_logger: OutputLogger
+    log: structlog.stdlib.BoundLogger,
+    output_logger: OutputLogger,
 ) -> t.AsyncGenerator[None, None]:
     meltano_stdout = output_logger.out(
         "meltano",