+++ src/meltano/cli/run.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 import typing as t
+import uuid
 
 import click
 import structlog
@@ -14,6 +15,7 @@ from meltano.cli.params import (
     pass_project,
 )
 from meltano.cli.utils import CliEnvironmentBehavior, CliError, PartialInstrumentedCmd
+from meltano.core._state import StateStrategy
 from meltano.core.block.block_parser import BlockParser, validate_block_sets
 from meltano.core.block.blockset import BlockSet
 from meltano.core.block.plugin_command import PluginCommandBlock
@@ -27,8 +29,6 @@ from meltano.core.tracking.contexts.plugins import PluginsTrackingContext
 from meltano.core.utils import run_async
 
 if t.TYPE_CHECKING:
-    import uuid
-
     from meltano.core.project import Project
 
 logger = structlog.getLogger(__name__)
@@ -85,6 +85,15 @@ install, no_install, only_install = get_install_options(include_only_install=Tru
     "--merge-state",
     is_flag=True,
     help="Merges state with that of previous runs.",
+    hidden=True,
+)
+@click.option(
+    "--state-strategy",
+    # TODO: use click.Choice(StateStrategy) once we drop support for Python 3.9 and use
+    # click 8.2+ exclusively
+    type=click.Choice([strategy.value for strategy in StateStrategy]),
+    default=StateStrategy.AUTO.value,
+    help="Strategy to use for state updates.",
 )
 @click.option(
     "--run-id",
@@ -112,6 +121,7 @@ async def run(
     force: bool,
     state_id_suffix: str,
     merge_state: bool,
+    state_strategy: str,
     run_id: uuid.UUID | None,
     blocks: list[str],
     install_plugins: InstallPlugins,
@@ -143,7 +153,16 @@ async def run(
         logger.info("Setting 'console' handler log level to 'debug' for dry run")
         change_console_log_level()
 
+    # Bind run_id at the start of the CLI entrypoint
+    run_id = run_id or uuid.uuid4()
+    structlog.contextvars.bind_contextvars(run_id=str(run_id))
+
     tracker: Tracker = ctx.obj["tracker"]
+
+    _state_strategy = StateStrategy.from_cli_args(
+        merge_state=merge_state,
+        state_strategy=state_strategy,
+    )
 
     try:
         parser = BlockParser(
@@ -155,7 +174,7 @@ async def run(
             no_state_update=no_state_update,
             force=force,
             state_id_suffix=state_id_suffix,
-            merge_state=merge_state,
+            state_strategy=_state_strategy,
             run_id=run_id,
         )
         parsed_blocks = list(parser.find_blocks(0))