+++ src/meltano/migrations/versions/990c0665f3ce_ensure_user_login_count_default_value.py
@@ -25,7 +25,7 @@ def upgrade() -> None:
     connection = op.get_bind()
 
     connection.execute(
-        user.update().where(user.c.login_count is None).values({"login_count": 0})
+        user.update().where(user.c.login_count is None).values({"login_count": 0}),
     )