+++ tests/fixtures/db/__init__.py
@@ -44,7 +44,7 @@ def vacuum_db(engine_sessionmaker):
     try:
         yield
     finally:
-        logging.debug("Cleaning system database...")  # noqa: TID251
+        logging.debug("Cleaning system database...")
         engine, _ = engine_sessionmaker
         close_all_sessions()
         metadata = MetaData()